Added some operations

+++ GroupProject/src/PropertyButton.java
@@ -11,7 +11,7 @@ public class PropertyButton extends Button implements PropertyButtonInterface{
 	private int level;
 	//private int time;
 	
-	public PropertyButton(int x, int y, int w, int h, String text, Color color, Action action, int price, int payout) {
+	public PropertyButton(int x, int y, int w, int h, String text, Color color, Action action, double price, double payout) {
 		super(x, y, w, h, text, color, action);
 		this.price = price;
 		this.payout = payout;
